Change wording slightly
This PR changes the wording to make clear that if you "use your Shelf object as a dictionary", the data is in fact _not_ "automatically saved in the file". I ended up making another minor tweak which I view as a slight improvement, but feel free to change it back if you do not like it.